+++ packages/react-components/src/components/cesium-map/layers/wms.layer.tsx
@@ -1,7 +1,7 @@
-import React from 'react';
+import React, { useMemo } from 'react';
 import { WebMapServiceImageryProvider } from 'cesium';
 import { CustomWebMapServiceImageryProvider } from '../helpers/customImageryProviders';
-import { useCesiumMap } from '../map';
+import { useCesiumMap, useCesiumMapViewstate } from '../map';
 import { CesiumImageryLayer, RCesiumImageryLayerProps } from './imagery.layer';
 
 export interface RCesiumWMSLayerOptions extends WebMapServiceImageryProvider.ConstructorOptions {}
@@ -13,10 +13,13 @@ export interface RCesiumWMSLayerProps extends Partial<RCesiumImageryLayerProps> 
 export const CesiumWMSLayer: React.FC<RCesiumWMSLayerProps> = (props) => {
   const { options, ...restProps } = props;
   const mapViewer = useCesiumMap();
+  const { viewState } = useCesiumMapViewstate();
 
-  const providerInstance = mapViewer.shouldOptimizedTileRequests
-    ? new CustomWebMapServiceImageryProvider(options, mapViewer)
-    : new WebMapServiceImageryProvider(options);
+  const providerInstance = useMemo(() => {
+    return viewState.shouldOptimizedTileRequests
+      ? new CustomWebMapServiceImageryProvider(options, mapViewer)
+      : new WebMapServiceImageryProvider(options);
+  }, [viewState.shouldOptimizedTileRequests]);
 
   return <CesiumImageryLayer {...restProps} imageryProvider={providerInstance} />;
 };