+++ packages/react-components/src/components/cesium-map/layers/wmts.layer.tsx
@@ -1,8 +1,8 @@
-import React from 'react';
+import React, { useMemo } from 'react';
 
 import { WebMapTileServiceImageryProvider } from 'cesium';
 import { CustomWebMapTileServiceImageryProvider } from '../helpers/customImageryProviders';
-import { useCesiumMap } from '../map';
+import { useCesiumMap, useCesiumMapViewstate } from '../map';
 import { CesiumImageryLayer, RCesiumImageryLayerProps } from './imagery.layer';
 
 export interface RCesiumWMTSLayerOptions extends WebMapTileServiceImageryProvider.ConstructorOptions {}
@@ -14,10 +14,13 @@ export interface RCesiumWMTSLayerProps extends Partial<RCesiumImageryLayerProps>
 export const CesiumWMTSLayer: React.FC<RCesiumWMTSLayerProps> = (props) => {
   const { options, ...restProps } = props;
   const mapViewer = useCesiumMap();
+  const { viewState } = useCesiumMapViewstate();
 
-  const providerInstance = mapViewer.shouldOptimizedTileRequests
-    ? new CustomWebMapTileServiceImageryProvider(options, mapViewer)
-    : new WebMapTileServiceImageryProvider(options);
+  const providerInstance = useMemo(() => {
+    return viewState.shouldOptimizedTileRequests
+      ? new CustomWebMapTileServiceImageryProvider(options, mapViewer)
+      : new WebMapTileServiceImageryProvider(options);
+  }, [viewState.shouldOptimizedTileRequests]);
 
   return <CesiumImageryLayer {...restProps} imageryProvider={providerInstance} />;
 };